Refactors GET /api/v1/foods endpoint and tests
