+++ routes/api/v1/foods.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const FoodsController = require('../../../controllers/foods-controller');
+
+router.get('/', FoodsController.index);
+
+module.exports = router;
