+++ test/foods.spec.js
@@ -1,0 +1,50 @@
+process.env.NODE_ENV = 'test';
+
+const chai = require('chai');
+const should = chai.should();
+const chaiHttp = require('chai-http');
+const server = require('../server');
+var knex = require('../db/knex');
+
+chai.use(chaiHttp);
+
+describe("API GET Food Endpoints", () => {
+  beforeEach((done) => {
+    knex.migrate.latest()
+    .then(() => {
+      knex.seed.run()
+      .then(() => {
+        done();
+      });
+    });
+  });
+
+  afterEach((done) => {
+    knex.migrate.rollback()
+    .then(() => {
+      done();
+    });
+  });
+
+  describe('GET /api/v1/foods', () => {
+    it('should return all of the foods if database contains foods', done => {
+      chai.request(server)
+      .get('/api/v1/foods')
+      .end((err, response) => {
+        response.should.have.status(200);
+        response.should.be.json;
+        response.body.should.be.a('array');
+        response.body.length.should.equal(2);
+        response.body[0].should.have.property('id');
+        response.body[0].id.should.equal(1);
+        response.body[0].should.have.property('name');
+        response.body[0].name.should.equal('Banana');
+        response.body[0].should.have.property('calories');
+        response.body[0].calories.should.equal(60);
+        response.body[0].should.not.have.property('created_at');
+        response.body[0].should.not.have.property('updated_at');
+        done();
+      });
+    });
+  });
+});
